red-whisperer/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0

# Testing
pytest==7.4.3
pytest-bdd==7.1.1
pytest-cov==4.1.0

# Security
cryptography==41.0.5
python-jose==3.3.0
passlib==1.7.4

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# Utilities
tenacity==8.2.3
aiohttp==3.9.1 